ASM-8072 Increase vcenter inventory timeout to 1 hour
In 8.3.1 the timeout was increased from 4 minutes to 15 minutes. The
script was timing out against some vcenters with large inventory
including remote ESXi hosts. Recently the same vcenter was seen to
take 24 minutes to inventory so it was still timing out. For now
bumping the timeout up to 1 hour to cover that case. At some point we
need to find some way to optimize this.